+++ RELEASE_NOTES.md
@@ -152,3 +152,5 @@ _(**NOT BACKWARDS COMPATIBLE**)_
 _(**NOT BACKWARDS COMPATIBLE**)_
 * Fixed getConnection.isOpen() still true after serial-port removal #74
 * Implemented inter-frame delays for serial transactions to match the Modbus specification #62
+* Fixed - Why doesn't DigitalOut extend DigitalIn #86
+